+++ Assets/Scripts/CharacterHealth.cs
@@ -152,7 +152,7 @@ public class CharacterHealth : MonoBehaviour, Health
         baseMaxHealth += (baseMaxHealth * 0.01f) * ((100 - level) * 0.1f);
         health += (baseMaxHealth * 0.01f) * ((100 - level) * 0.1f);
         maxHealth += (baseMaxHealth * 0.01f) * ((100 - level) * 0.1f);
-        GetComponent<Character>().Health.SetBaseValue(baseMaxHealth);
+        GetComponent<Character>().Health.SetBaseValue((int)baseMaxHealth);
         GetComponent<Character>().statPanel.UpdateStatValues();
     }
